+++ src/main/java/module-info.java
@@ -1,5 +1,5 @@
 import org.cryptomator.integrations.mount.MountService;
-import org.cryptomator.integrations.revealfiles.RevealPathsService;
+import org.cryptomator.integrations.revealpaths.RevealPathsService;
 import org.cryptomator.integrations.tray.TrayMenuController;
 import org.cryptomator.integrations.autostart.AutoStartProvider;
 import org.cryptomator.integrations.keychain.KeychainAccessProvider;
@@ -15,7 +15,7 @@ module org.cryptomator.integrations.api {
 	exports org.cryptomator.integrations.common;
 	exports org.cryptomator.integrations.keychain;
 	exports org.cryptomator.integrations.mount;
-	exports org.cryptomator.integrations.revealfiles;
+	exports org.cryptomator.integrations.revealpaths;
 	exports org.cryptomator.integrations.tray;
 	exports org.cryptomator.integrations.uiappearance;
